Move risc-v driver unit-test to a new file. Separate it into two test cases
