+++ simulator/risc_v/t/unit_test.cpp
@@ -4,20 +4,11 @@
  * Copyright 2019 MIPT-MIPS
  */
 
-#include <risc_v/risc_v.h>
-#include <simulator.h>
 #include "riscv_test_wrapper.h"
 
 #include <catch.hpp>
 #include <memory/memory.h>
 #include <risc_v/riscv_instr.h>
-
-static auto get_op_with_trap( Trap trap)
-{
-    Operation op( 0x100, 0x104);
-    op.set_trap( trap);
-    return op;
-}
 
 TEST_CASE("RISCV disassembly")
 {
@@ -610,16 +601,4 @@ TEST_CASE("RISCV ecall")
     RISCVInstr<uint32> instr( "ecall", 0);
     instr.execute();
     CHECK( instr.trap_type() == Trap::SYSCALL);
-}
-
-TEST_CASE("RISCV32 driver")
-{
-    auto sim = Simulator::create_configured_isa_simulator( "riscv32");
-    auto drv = create_riscv32_driver( sim.get());
-    CHECK( drv->handle_trap( get_op_with_trap( Trap( Trap::NO_TRAP))) == Trap::NO_TRAP);
-    CHECK( drv->handle_trap( get_op_with_trap( Trap( Trap::HALT))) == Trap::HALT);
-    auto expected_pc = (sim->read_csr_register( "stvec") >> 2U) & ~(bitmask<uint64>( 3));
-    CHECK( sim->get_pc() == expected_pc);
-    auto expected_cause = Trap( Trap::HALT).to_riscv_format();
-    CHECK( sim->read_csr_register( "scause") == expected_cause);
 }
